refactor: move wallet canister creation methods to wallet.rs
