+++ src/dfx/src/commands/canister/create.rs
@@ -5,7 +5,6 @@ use crate::lib::ic_attributes::{
 };
 use crate::lib::identity::identity_manager::IdentityManager;
 use crate::lib::identity::identity_utils::CallSender;
-use crate::lib::identity::Identity;
 use crate::lib::operations::canister::create_canister;
 use crate::lib::root_key::fetch_root_key_if_needed;
 use crate::util::clap::validators::cycle_amount_validator;
@@ -13,6 +12,7 @@ use crate::util::clap::validators::{
     compute_allocation_validator, freezing_threshold_validator, memory_allocation_validator,
 };
 
+use crate::lib::identity::wallet::get_or_create_wallet_canister;
 use anyhow::{anyhow, bail, Context};
 use candid::Principal as CanisterId;
 use clap::Parser;
@@ -76,7 +76,7 @@ pub async fn exec(
 
     let proxy_sender;
     if !opts.no_wallet && !matches!(call_sender, CallSender::Wallet(_)) {
-        let wallet = Identity::get_or_create_wallet_canister(
+        let wallet = get_or_create_wallet_canister(
             env,
             env.get_network_descriptor(),
             env.get_selected_identity().expect("No selected identity"),
