+++ src/dfx/src/lib/identity/mod.rs
@@ -8,23 +8,19 @@ use crate::lib::environment::Environment;
 use crate::lib::error::{DfxResult, IdentityError};
 use crate::lib::identity::identity_manager::IdentityStorageMode;
 use crate::lib::network::network_descriptor::{NetworkDescriptor, NetworkTypeDescriptor};
-use crate::lib::root_key::fetch_root_key_if_needed;
 use dfx_core::error::wallet_config::WalletConfigError;
 use dfx_core::error::wallet_config::WalletConfigError::{
     EnsureWalletConfigDirFailed, LoadWalletConfigFailed, SaveWalletConfigFailed,
 };
 use dfx_core::json::{load_json_file, save_json_file};
 
-use anyhow::{anyhow, bail, Context};
+use anyhow::{bail, Context};
 use bip39::{Language, Mnemonic};
 use candid::Principal;
 use fn_error_context::context;
 use ic_agent::identity::{AnonymousIdentity, BasicIdentity, Secp256k1Identity};
-use ic_agent::{AgentError, Signature};
+use ic_agent::Signature;
 use ic_identity_hsm::HardwareIdentity;
-use ic_utils::call::AsyncCall;
-use ic_utils::interfaces::management_canister::builders::InstallMode;
-use ic_utils::interfaces::{ManagementCanister, WalletCanister};
 use sec1::EncodeEcPrivateKey;
 use serde::{Deserialize, Serialize};
 use slog::{debug, info, trace, Logger};
@@ -35,13 +31,12 @@ pub mod identity_manager;
 pub mod identity_utils;
 pub mod keyring_mock;
 pub mod pem_safekeeping;
-use crate::util::assets::wallet_wasm;
+pub mod wallet;
+
 pub use identity_manager::{
     HardwareIdentityConfiguration, IdentityConfiguration, IdentityCreationParameters,
     IdentityManager,
 };
-
-use super::diagnosis::DiagnosedError;
 
 pub const ANONYMOUS_IDENTITY_NAME: &str = "anonymous";
 pub const IDENTITY_PEM: &str = "identity.pem";
@@ -487,104 +482,6 @@ impl Identity {
         Ok(())
     }
 
-    #[context("Failed to create wallet for identity '{}' on network '{}'.", name, network.name)]
-    pub async fn create_wallet(
-        env: &dyn Environment,
-        network: &NetworkDescriptor,
-        name: &str,
-        some_canister_id: Option<Principal>,
-    ) -> DfxResult<Principal> {
-        fetch_root_key_if_needed(env).await?;
-        let mgr = ManagementCanister::create(
-            env.get_agent()
-                .ok_or_else(|| anyhow!("Cannot get HTTP client from environment."))?,
-        );
-        info!(
-            env.get_logger(),
-            "Creating a wallet canister on the {} network.", network.name
-        );
-
-        let wasm = wallet_wasm(env.get_logger())?;
-
-        let canister_id = match some_canister_id {
-            Some(id) => id,
-            None => {
-                mgr.create_canister()
-                    .as_provisional_create_with_amount(None)
-                    .with_effective_canister_id(env.get_effective_canister_id())
-                    .call_and_wait()
-                    .await
-                    .context("Failed create canister call.")?
-                    .0
-            }
-        };
-
-        match mgr
-            .install_code(&canister_id, wasm.as_slice())
-            .with_mode(InstallMode::Install)
-            .call_and_wait()
-            .await
-        {
-            Err(AgentError::ReplicaError {
-                reject_code: 5,
-                reject_message,
-            }) if reject_message.contains("not empty") => {
-                bail!(
-                    r#"The wallet canister "{canister_id}" already exists for user "{name}" on "{}" network."#,
-                    network.name
-                )
-            }
-            res => res.context("Failed while installing wasm.")?,
-        }
-
-        let wallet = Identity::build_wallet_canister(canister_id, env).await?;
-
-        wallet
-            .wallet_store_wallet_wasm(wasm)
-            .call_and_wait()
-            .await
-            .context("Failed to store wallet wasm.")?;
-
-        Identity::set_wallet_id(network, name, canister_id)?;
-
-        info!(
-            env.get_logger(),
-            r#"The wallet canister on the "{}" network for user "{}" is "{}""#,
-            network.name,
-            name,
-            canister_id,
-        );
-
-        Ok(canister_id)
-    }
-
-    /// Gets the currently configured wallet canister. If none exists yet and `create` is true, then this creates a new wallet. WARNING: Creating a new wallet costs ICP!
-    ///
-    /// While developing locally, this always creates a new wallet, even if `create` is false.
-    /// This can be inhibited by setting the DFX_DISABLE_AUTO_WALLET env var.
-    #[context("Failed to get wallet for identity '{}' on network '{}'.", name, network.name)]
-    pub async fn get_or_create_wallet(
-        env: &dyn Environment,
-        network: &NetworkDescriptor,
-        name: &str,
-    ) -> DfxResult<Principal> {
-        match Identity::wallet_canister_id(network, name)? {
-            None => {
-                // If the network is not the IC, we ignore the error and create a new wallet for the identity.
-                if !network.is_ic && std::env::var("DFX_DISABLE_AUTO_WALLET").is_err() {
-                    Identity::create_wallet(env, network, name, None).await
-                } else {
-                    Err(DiagnosedError::new(format!("This command requires a configured wallet, but the combination of identity '{}' and network '{}' has no wallet set.", name, network.name),
-                    "To use an identity with a configured wallet you can do one of the following:\n\
-                    - Run the command for a network where you have a wallet configured. To do so, add '--network <network name>' to your command.\n\
-                    - Switch to an identity that has a wallet configured using 'dfx identity use <identity name>'.\n\
-                    - Configure a wallet for this identity/network combination: 'dfx identity set-wallet <wallet id> --network <network name>'.".to_string())).context("Wallet not configured.")
-                }
-            }
-            Some(principal) => Ok(principal),
-        }
-    }
-
     #[context("Failed to get wallet canister id for identity '{}' on network '{}'.", name, network.name)]
     pub fn wallet_canister_id(
         network: &NetworkDescriptor,
@@ -602,43 +499,6 @@ impl Identity {
             .get(name)
             .and_then(|wallet_network| wallet_network.networks.get(&network.name).cloned());
         Ok(maybe_wallet_principal)
-    }
-
-    #[context("Failed to construct wallet canister caller.")]
-    pub async fn build_wallet_canister(
-        id: Principal,
-        env: &dyn Environment,
-    ) -> DfxResult<WalletCanister<'_>> {
-        Ok(WalletCanister::from_canister(
-            ic_utils::Canister::builder()
-                .with_agent(
-                    env.get_agent()
-                        .ok_or_else(|| anyhow!("Cannot get HTTP client from environment."))?,
-                )
-                .with_canister_id(id)
-                .build()
-                .unwrap(),
-        )
-        .await?)
-    }
-
-    /// Gets the currently configured wallet canister. If none exists yet and `create` is true, then this creates a new wallet. WARNING: Creating a new wallet costs ICP!
-    ///
-    /// While developing locally, this always creates a new wallet, even if `create` is false.
-    /// This can be inhibited by setting the DFX_DISABLE_AUTO_WALLET env var.
-    #[allow(clippy::needless_lifetimes)]
-    #[context("Failed to get wallet canister caller for identity '{}' on network '{}'.", name, network.name)]
-    pub async fn get_or_create_wallet_canister<'env>(
-        env: &'env dyn Environment,
-        network: &NetworkDescriptor,
-        name: &str,
-    ) -> DfxResult<WalletCanister<'env>> {
-        // without this async block, #[context] gives a spurious error
-        async {
-            let wallet_canister_id = Identity::get_or_create_wallet(env, network, name).await?;
-            Identity::build_wallet_canister(wallet_canister_id, env).await
-        }
-        .await
     }
 }
