+++ src/dfx/src/lib/operations/canister/install_canister.rs
@@ -3,7 +3,6 @@ use crate::lib::canister_info::CanisterInfo;
 use crate::lib::environment::Environment;
 use crate::lib::error::DfxResult;
 use crate::lib::identity::identity_utils::CallSender;
-use crate::lib::identity::Identity;
 use crate::lib::installers::assets::post_install_store_assets;
 use crate::lib::models::canister::CanisterPool;
 use crate::lib::models::canister_id_store::CanisterIdStore;
@@ -12,6 +11,7 @@ use crate::lib::network::network_descriptor::NetworkDescriptor;
 use crate::util::assets::wallet_wasm;
 use crate::util::read_module_metadata;
 
+use crate::lib::identity::wallet::build_wallet_canister;
 use anyhow::{anyhow, bail, Context};
 use backoff::backoff::Backoff;
 use backoff::ExponentialBackoff;
@@ -182,7 +182,7 @@ pub async fn install_canister(
     }
     if canister_info.is_assets() {
         if let CallSender::Wallet(wallet_id) = call_sender {
-            let wallet = Identity::build_wallet_canister(*wallet_id, env).await?;
+            let wallet = build_wallet_canister(*wallet_id, env).await?;
             let identity_name = env.get_selected_identity().expect("No selected identity.");
             info!(
                 log,
@@ -397,7 +397,7 @@ YOU WILL LOSE ALL DATA IN THE CANISTER.");
                 .context("Failed to install wasm.")?;
         }
         CallSender::Wallet(wallet_id) => {
-            let wallet = Identity::build_wallet_canister(*wallet_id, env).await?;
+            let wallet = build_wallet_canister(*wallet_id, env).await?;
             let install_args = CanisterInstall {
                 mode,
                 canister_id,
@@ -432,7 +432,7 @@ pub async fn install_wallet(
         .call_and_wait()
         .await
         .context("Failed to install wallet wasm.")?;
-    let wallet = Identity::build_wallet_canister(id, env).await?;
+    let wallet = build_wallet_canister(id, env).await?;
     wallet
         .wallet_store_wallet_wasm(wasm)
         .call_and_wait()
